Moved bookmark managing code from ViewController to BookmarkManager.
In order to not clutter the ViewController class, bookmark managing was
moved to the BookmarkManager class.
